move i2c_common dir to hal top
