+++ hal/i2c_common.c
@@ -6,7 +6,7 @@
  */
 
 #include "i2c_common.h"
-#include "../../Library/endian.h"
+#include "../Library/endian.h"
 #include <string.h>
 
 /**
